// ChatWeb/chat-app/src/app/models/realModel/user.model.ts
export interface Userr {
    _id: string;
    name: string;
    avatarUrl?: string | 'https://i1.rgstatic.net/ii/profile.image/1039614412341248-1624874799001_Q512/Meryem-Laval.jpg';
    // online: boolean;
    // address?:string;
    email:string;
    // phone:string;
    // friends:string[];
  }
  